requirements.txt
svgwrite==1.4.3
Pillow==10.2.0
flask==3.0.2
flask-cors>=4.0.0
numpy>=1.21.0
tensorflow>=2.8.0
matplotlib>=3.5.0
requests>=2.28.0
mediapipe>=0.10.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
joblib>=1.0.0
opencv-python-headless